formalize routes.py
